+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Generated CSV files now include an alphabetically sorted list of all parameters at the end of the file, including the description and example text.
 
 ### Changed